Share the template property warning payload through a base class
The two warnings a rename fix binds to carried the same four members and the
same range logic, which SonarCloud flagged as 45 duplicated lines each. Move
that into TemplatePropertyWarningBase, next to the other *WarningBase types.
The severity attributes need a per-class const, so the message, the severity id
and the tooltip stay in the derived classes.

The fix base now assigns the payload once and its two constructors, which exist
only because a quick fix binds by constructor parameter type, chain into it.
