+++ src/ReSharper.Structured.Logging/Highlighting/PositionalPropertyUsedWarning.cs
@@ -1,5 +1,4 @@
 using JetBrains.Annotations;
-using JetBrains.DocumentModel;
 using JetBrains.ReSharper.Feature.Services.Daemon;
 using JetBrains.ReSharper.Psi.CSharp;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
@@ -22,7 +21,7 @@ namespace ReSharper.Structured.Logging.Highlighting
         CSharpLanguage.Name,
         OverlapResolve = OverlapResolveKind.WARNING,
         ToolTipFormatString = Message)]
-    public class PositionalPropertyUsedWarning : IHighlighting
+    public class PositionalPropertyUsedWarning : TemplatePropertyWarningBase, IHighlighting
     {
         private const string Message = "Prefer named properties instead of positional ones";
 
@@ -33,44 +32,12 @@ namespace ReSharper.Structured.Logging.Highlighting
             [NotNull] PropertyToken namedProperty,
             [CanBeNull] ICSharpArgument argument,
             [NotNull] string[] usedPropertyNames)
+            : base(tokenInformation, namedProperty, argument, usedPropertyNames)
         {
-            TokenInformation = tokenInformation;
-            NamedProperty = namedProperty;
-            Argument = argument;
-            UsedPropertyNames = usedPropertyNames;
         }
-
-        [NotNull]
-        public MessageTemplateTokenInformation TokenInformation { get; }
-
-        [NotNull]
-        public PropertyToken NamedProperty { get; }
-
-        /// <summary>
-        /// The argument that fills the hole, or <c>null</c> when there is none to derive a name from,
-        /// as in an attribute template or when the values are passed as a single array.
-        /// </summary>
-        [CanBeNull]
-        public ICSharpArgument Argument { get; }
-
-        /// <summary>
-        /// Every property name the template already uses, so a fix can avoid renaming into a collision.
-        /// </summary>
-        [NotNull]
-        public string[] UsedPropertyNames { get; }
 
         public string ErrorStripeToolTip => ToolTip;
 
         public string ToolTip => Message;
-
-        public DocumentRange CalculateRange()
-        {
-            return TokenInformation.DocumentRange;
-        }
-
-        public bool IsValid()
-        {
-            return TokenInformation.DocumentRange.IsValid();
-        }
     }
 }
